pyproject.toml
[tool.poetry]
name = "beside_project"
version = "0.1.0"
description = ""
authors = ["KimDuHong <art970@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
Django = "^4.2.4"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.1.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
sentry-sdk = "^1.29.2"
django-cors-headers = "^4.2.0"
gunicorn = "^21.2.0"
django-environ = "^0.10.0"
boto3 = "^1.28.20"
requests = "^2.31.0"
Pillow = "^10.0.0"
imageio = "^2.31.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
